Add CORS support for Refresh Token Grant.
